Add reserva habitacion

+++ services/reservas-svc/src/main/java/edu/utn/frsf/isi/dan/reservas_svc/model/Reserva.java
@@ -17,7 +17,7 @@ public class Reserva {
     @Id
     private String _id;
     private String idHabitacion;
-    private Long hotelId;
+    private Integer hotelId;
     private Instant createdAt;
     private Instant checkIn;
     private Instant checkOut;
